+++ backend/source/schemas/discounts_s.py
@@ -8,7 +8,7 @@ class UpdateDiscountRequest(BaseModel):
     model_config = ConfigDict(extra="forbid")
     name: str | None = None
     type: Literal["percent", "fixed"] | None = None
-    value: float | None = Field(default=None, gt=0)
+    value: int | None = Field(default=None, gt=0)
     scope: Literal["all", "category", "product", "product_list"] | None = None
     scope_value: str | None = None
     is_active: bool | None = None
@@ -21,7 +21,7 @@ class CreateDiscountRequest(BaseModel):
     model_config = ConfigDict(extra="forbid")
     name: str
     type: Literal["percent", "fixed"]
-    value: float = Field(gt=0)
+    value: int = Field(gt=0)
     scope: Literal["all", "category", "product", "product_list"]
     scope_value: str | None = None
     is_active: bool = True
